sample: swap -ss option to before the input file
Makes ffmpeg seek instantly to the point in time of the video file the
sample is to be taken from; using it as an output option instead makes
the video decode from the beginning up until that point.

Saving 40 minutes of processing time seems typical for some of my
Blu-ray files.